+++ processor.py
@@ -1252,7 +1252,9 @@ def handle_create_component(data, component_name, component_type, file_name, tim
         # Save settings and show success message
         sublime.save_settings(COMPONENT_METADATA_SETTINGS)
         file_base_name = component_name + extension
-        print (message.SEPRATE.format("{0} is created successfully".format(file_base_name)))
+        print (message.SEPRATE.format(
+            "{0} is created successfully at {1}".format(file_base_name, 
+                time.strftime("%Y-%m-%d %H:%M:%S", time.localtime(time.time())))))
                 
     toolingapi_settings = context.get_toolingapi_settings()
     api = SalesforceApi(toolingapi_settings)
@@ -1325,6 +1327,10 @@ def handle_delete_component(component_url, file_name, timeout=120):
         if result["status_code"] > 399: return
         os.remove(file_name)
         sublime.active_window().run_command("close")
+
+        print (message.SEPRATE.format(
+            "{0} is deleted successfully at {1}".format(file_name, 
+                time.strftime("%Y-%m-%d %H:%M:%S", time.localtime(time.time())))))
 
     toolingapi_settings = context.get_toolingapi_settings()
     api = SalesforceApi(toolingapi_settings)